rename server to run
