+++ discourse_cleanup.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-#
+
 # # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
